base de datos PostgreSql

+++ Urbans_Sneakers/Productos/forms.py
@@ -15,7 +15,7 @@ class Formulario_Product(forms.Form):
     category=forms.ModelChoiceField(label='Categoria', queryset=Category.objects.all())
     sub_category=forms.ModelChoiceField(label='Sub-Categoria', queryset=Sub_Category.objects.all())
     model=forms.CharField(max_length=40)
-    price=forms.IntegerField()
+    price=forms.FloatField()
     coulor=forms.CharField(max_length=40,required=False)
     description=forms.CharField(max_length=400)
     stock=forms.IntegerField()
